modification après retour

+++ server/src/main/java/fr/uga/l3miage/tp1/exo2/models/NetflixAccountEntity.java
@@ -1,14 +1,13 @@
 package fr.uga.l3miage.tp1.exo2.models;
 
 import javax.persistence.*;
-import java.math.BigInteger;
-import fr.uga.l3miage.tp1.exo2.models.NetflixUserEntity;
+
 
 @Entity(name = "netflix_account")
 public class NetflixAccountEntity {
 @Id
     @Column(name = "id")
-    private BigInteger id;
+    private long id;
 @Column(name = "nb_devices")
     private Integer nbDevices;
 @Column(name = "type_account")
